+++ package.json
@@ -7,6 +7,7 @@
     "firebase": "^9.6.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-firebase-hooks": "^4.0.2",
     "react-router-dom": "^6.2.1",
     "react-scripts": "4.0.1"
   },
